android: refactor log
So it can be accessed everywhere
