--- android/Lib/src/main/kotlin/com/darakeon/dfm/lib/extensions/log.kt
@@ -1,11 +1,0 @@
-package com.darakeon.dfm.lib.extensions
-
-import android.util.Log
-
-fun log(text: Any) {
-	try {
-		Log.e("DFM", text.toString())
-	} catch(e: Exception) {
-		println(text)
-	}
-}
